Fix integer overflow in Parent Access Code authenticator
Large timestamps (beyond the year 6055) caused the 32-bit loop counter
in Authenticator::ValidateInRange to wrap around, leading to a
brute-force vulnerability. Changing the counter type to int64_t prevents
this.

Fixed: 513606986
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7875246
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1637792}

+++ chrome/browser/ash/child_accounts/parent_access_code/authenticator.cc
@@ -166,7 +166,7 @@ std::optional<AccessCode> Authenticator::ValidateInRange(
                                  kAccessCodeGranularity.InMilliseconds();
   const int64_t end_interval = valid_to.InMillisecondsSinceUnixEpoch() /
                                kAccessCodeGranularity.InMilliseconds();
-  for (int i = start_interval; i <= end_interval; ++i) {
+  for (int64_t i = start_interval; i <= end_interval; ++i) {
     const base::Time generation_timestamp =
         base::Time::FromMillisecondsSinceUnixEpoch(
             i * kAccessCodeGranularity.InMilliseconds());
